Adding some abstractions to hide away the Hoodie implementation details.
`#checkpoint` encapsulated commit management APIs
`#wip` encapsulating reading of commits.

``#todo` clarify the use of `UPSERT` with `Append`

Introduced abstraction for mapping from raw to Hoodie format.

Introduced helper method to `getIds` and logging to verify ingestion uses the right data.

`#todo` was there a bug on line 186 - using ._2 instead of ._3 or am I missing something?

`#fix` performances are not ingested correctly due to how the raw files are loaded and partitioned.
Changed `getPerformances_3Split` to split each data set in 3 and then merge the chunks so that each data set is guaranteed to be split and thus participate in the incremental updates.

Small refactoring to clarify and improve performance.

Introduced `spark-daria` for custom transformations.
Wip on an error.